comma and dot

+++ app/model/main_base.py
@@ -6,7 +6,7 @@ from app.model.spravichnik import Spravichnik
 class Main(models.Model):
     rasxod = models.ForeignKey(Spravichnik, on_delete=models.CASCADE, related_name='rasxod', null=True)
     prixod = models.ForeignKey(Spravichnik, on_delete=models.CASCADE, related_name='prixod')
-    summa = models.IntegerField(blank=True, null=True)
+    summa = models.FloatField(blank=True, null=True)
     date_main = models.DateField()
     provodka = models.TextField(null=True, blank=True)
     active = models.BooleanField(default=True)
